Compute Attack Rates

For dts models, compute the attack rates, the expected probability of infection over one time step, is a weighted sum of local and travel exposure.

# S3 method for class 'dts'
Exposure(t, y, pars)

Arguments

t

the time

y

the state variables

pars

an xds object

Value

an xds object

Details

Local attack rates (AR or \(\alpha\)) are a function of local daily entomological inoculation rates (dEIR or \(E\)), and a model for the probability of infection per infectious bite, \(b\) (from \(F_b\), as defined by an \(\cal X\) model). The total attack rates also consider exposure to malaria while traveling, (\(T_\delta\)): $$\alpha = (1-\delta) \; F_\alpha(E, b) + \delta\; T_\alpha(b)$$

See also