Compute and store availability of egg-laying habitats

Set the value of a variable, \(Q\), that describes the availability of any habitat or device that would attract mosquitoes and induce them to lay eggs.

Usage

make_Q(pars)

Arguments

pars

an xds object

Value

an xds object

See also