Setup Egg Laying

Set up a part of the xds object that defines the interface for egg laying

Usage

setup_EGG_LAYING(pars, membership)

Arguments

pars

an xds object

membership

is the habitat membership vector

Value

an xds object

Details

This implements a model for egg laying described by Wu SL, et al., (2023).

Modular computation in ramp.xds requires a rigid interface to guarantee mathematical consistency for egg laying and emergence. The interface is defined by an object called EGGpar that is attached to the xds object pars as pars$EGGpar. The interface includes

a habitat membership matrix, \(\cal N\) made by create_habitat_matrix

the habitat search weights

a quantity that is motivated by mosquito searching for resources, called habitat availability \(Q\) made by compute_Q;

the availability of ovitraps

the availability of unsuitable habitats

the availability of anything that is like a habitat, but that is an aquatic habitat in the model, including ovitraps and unsuitable habitats, \(Q_{tot}\) made by compute_Qtot;

the egg distribution matrix \(\cal U\), made by compute_calU

a vector that stores eggs laid

This function is called by make_xds_template to set up EGGpar and the variables and parameters with all the variables it might depend on.

References

WuSL2023SpatialDynamicsramp.xds

See also