ramp.xds: Getting Help

ramp.xds concepts and functions are well documented. These help guides are designed to find what you need.

xds_help_basic_setup is an overview of basic setup

xds_help_setup_options is an overview of other options

get_ functions inspect a model component or outputs

change_ functions modify part of a model component

setup_ functions replace a model component

XH Component

Type ?XH_ to find a list of help guides for human / host demography, infection, & immunity.

XH_module_list – list of XH Modules

XH_functions – list of functions for XH Modules

XH_get – list of get_ functions for XH Modules

XH_change – list of change_ functions for XH Modules

XH_setup – list of setup_ functions for XH Modules

MY Component

Type ?MY_ to find a list of help guides for adult mosquito ecology & infection dynamics.

MY_module_list – list of MY Modules

MY_functions – list of functions for MY Modules

MY_get – list of get_ functions for MY Modules

MY_change – list of change_ functions for MY Modules

MY_setup – list of setup_ functions for MY Modules

L Component

Type ?L_ to find a list of help guides for immature mosquito ecology.

L_module_list – list of L Modules

L_functions – list of functions for L Modules

L_get – list of get_ functions for L Modules

L_change – list of change_ functions for L Modules

L_setup – list of setup_ functions for L Modules