Build a Model and Configure All Components

Make an xds model object:

Define a module for each dynamical component (see dynamical_components):

XH Component – human / host infection dynamics

MY Component – adult mosquito ecology and infection dynamics

L Component – aquatic mosquito ecology

Define basic structural parameters for a single host and vector population:

nPatches - the number of patches (see patch_dynamics)

membership - the habitat membership vector (see aquatic_habitats)

nHabitats = length(membership)

residence - the human residence vector (see human_populations)

HPop - the human population size (see human_populations)

nStrata = length(residence) = length(HPop)

Configure some of the basic elements:

Search Weights (see search_weights)

Mosquito Dispersal matrix (if nPatches>1; see mosquito_dispersal)

Time Spent matrix (if nPatches>1; see time_spent)

Usage

xds_setup(
  xds = "ode",
  Xname = "SIS",
  XHoptions = list(),
  MYname = "SI",
  MYoptions = list(),
  Lname = "trivial",
  Loptions = list(),
  nPatches = 1,
  HPop = 1000,
  residence = 1,
  TimeSpent = list(),
  searchB = 1,
  membership = 1,
  Koptions = list(Kname = "no_setup"),
  searchQ = 1,
  BFopts = list(),
  model_name = "unnamed"
)

Arguments

xds

is ode or dde or dts for ordinary OR delay differential OR difference equations

Xname

a character string defining a X Component module

XHoptions

a list to configure the X Component module

MYname

a character string defining an MY module

MYoptions

options to set up the MY component

Lname

a character string defining a L Component module

Loptions

a list to configure the L Component module

nPatches

is the number of patches

HPop

is the number of humans in each patch

residence

is a vector that describes the patch where each human stratum lives

TimeSpent

is either a TimeSpent matrix or a string to call a function that sets it up

searchB

is a vector of search weights for blood feeding

membership

is a vector that describes the patch where each aquatic habitat is found

Koptions

a K matrix, or options for setup_K_matrix (see mosquito_dispersal)

searchQ

is a vector of search weights for egg laying

BFopts

a list to configure the blood feeding model

model_name

is a name for the model (arbitrary)

Value

an xds object

Details

make_xds_object_template returns an object template with a properly configured interface for blood feeding and egg laying (see xds_object):

nPatches is passed as a parameter

nHabitats = length(membership)

nStrata = length(residence) = length(HPop)

nHostSpecies=1 (basic setup handles only the first host species)

nVectorSpecies=1 (basic setup handles only the first vector species)

class(xds_obj$frame) = 'full'

Each one of the dynamical components is configured.

xds_obj$XH_obj[[1]] defines a model for human / host infection dynamics of class Xname (the X component). The parameter values passed in a named list, XHoptions.

xds_obj$MY_obj[[1]] defines a model for adult mosquito ecology & infection dynamics of class MYname (the MY Component). The parameter values are passed in a named list, MYoptions.

xds_obj$L_obj[[1]] defines a model for aquatic mosquito ecology of class Lname (The L Component). The parameter values are passed in a named list, Loptions.

After configuring the dynamical components, several structural parameters can be configured at the command line:

Habitat search weights can be set

Host population search weights can be set

A mosquito dispersal matrix, \(\cal Koptions\), can be set

A time spent matrix, \(\Theta\), can be set

Note

Other options can be configured after basic setup (see xds_help_setup_options). If the MY Component is the trivial module, consider using xds_setup_aquatic or xds_setup_human or xds_setup_eir. Models for mosquito ecology need not include states describing infection status (see xds_setup_mosy).

See also