Make an xds Object to Study Mosquito Ecology

A modified version of xds_setup that streamlines setup for models that do not require a component describing parasite / pathogen infection dynamics in the adult mosquito population.

The xds object defines frame = class(frame) = 'mosy' to dispatch xde_derivatives.mosy or dts_update.mosy and associated functions.

The X Component module is trivial, but since humans / vertebrate hosts can be a resource, HPop must be set.

Usage

xds_setup_mosy(
  xds = "ode",
  MYZname = "basicM",
  Lname = "basicL",
  nPatches = 1,
  membership = 1,
  HPop = 1000,
  MYZday = 1,
  Lday = 1,
  calK = list(),
  searchQ = 1,
  kappa = 0,
  MYZopts = list(),
  Lopts = list(),
  model_name = "unnamed"
)

Arguments

xds

is ode or dde or dts for ordinary OR delay differential OR difference equations

MYZname

is a character string defining a MYZ Component module

Lname

is a character string defining a L Component module

nPatches

is the number of patches

membership

is a vector that describes the patch where each aquatic habitat is found

HPop

is the human population density

MYZday

is the run-time time step for MYZ Component (in days): integer or 1/integer

Lday

is the run-time time step for L Component (in days): integer or 1/integer

calK

is either a calK matrix or a string that defines how to set it up

searchQ

is a vector of search weights for egg laying

kappa

is a vector describing net infectiousness

MYZopts

a list to configure the MYZ Component module

Lopts

a list to configure the L Component module

model_name

is a name for the model (arbitrary)

Value

an xds object

See also