Derivatives for basicL (L Component)

This implements differential equation model for aquatic mosquito ecology. The equations have been modified slightly from a version published by Smith DL, et al. (2013); this version includes delayed maturation i in response to mean crowding.

Variables:

\(L\): the density of mosquito larvae in each habitat

Parameters and Terms:

\(\eta\) or eta: egg deposition rate (from F_eggs)

\(\psi\) or psi: maturation rate

\(\xi\) or xi: delayed maturation parameter in response to mean crowding

\(\phi\) or phi: density-independent death rate

\(\theta\) or theta: the slope of the mortality rate in response to mean crowding

Dynamics:

$$dL/dt = \eta - (\psi\;e^{-\xi L} + \phi + \theta L)L$$

Per-capita mortality is thus \(\phi + \theta L\), and the emergence rate of adult mosquitoes is \(\psi e^{-\xi L} L \)

# S3 method for basicL
dLdt(t, y, pars, s)

Arguments

t

current simulation time

y

state vector

pars

an xds object

s

the species index

Value

a numeric vector

References

Smith DL, Perkins TA, Tusting LS, Scott TW, Lindsay SW (2013). “Mosquito population regulation and larval source management in heterogeneous environments.” PLoS ONE, 8(8), e71247. doi:10.1371/journal.pone.0071247 .

See also