Forcing with Trivial Modules

Trivial modules were developed for each one of the three dynamical components, making it possible to develop studies of some focal process with known inputs: a trace function approach.

Three trivial modules and EIR forcing xds_setup_eir construct trace functions as decomposable time series. The value of a forced variable \(x(t)\) is computed as a product of four configurable elements:

\(\bar x\): a mean value

\(S(t)\): a seasonal pattern

\(T(t)\): a trend

\(K(t)\): a shock

$$x(t) = \bar x \times S(t) \times T(t) \times K(t)$$

In the trivial modules, the functions are specified by passing parameters generated for make_function.

season_par creates \(S(t)\) or F_season (eg, using makepar_F_sin)

trend_par creates \(T(t)\) or F_trend (eg, using makepar_F_spline)

shock_par creates \(K(t)\) or F_shock (eg, using makepar_F_sharkbite)

For a discussion of mechanistic approaches to forcing, see Forcing

Trivial modules were developed for each one of the three dynamical components, making it possible to develop studies of some focal process with known inputs: a trace function approach.

Three trivial modules and EIR forcing xds_setup_eir construct trace functions as decomposable time series. The value of a forced variable \(x(t)\) is computed as a product of four configurable elements:

\(\bar x\): a mean value

\(S(t)\): a seasonal pattern

\(T(t)\): a trend

\(K(t)\): a shock

$$x(t) = \bar x \times S(t) \times T(t) \times K(t)$$

In the trivial modules, the functions are specified by passing parameters generated for make_function.

season_par creates \(S(t)\) or F_season (eg, using makepar_F_sin)

trend_par creates \(T(t)\) or F_trend (eg, using makepar_F_spline)

shock_par creates \(K(t)\) or F_shock (eg, using makepar_F_sharkbite)

For a discussion of mechanistic approaches to forcing, see Forcing

See also