Build an xds object

Set up an xds object and set pars$frame <- class(pars$frame) <- 'full'.

Each dynamical component is fully configured for a single host and vector species:

Xpar defines a model for human / host infection dynamics of class Xname (the \(\cal XH\) dynamical component)

MYZpar defines a model for adult mosquito ecology & infection dynamics of class MYZname (the \(\cal MYZ\) dynamical component)

Lpar defines the aquatic mosquito ecology of class Lname (The \(\cal L\) dynamical component)

The model sets the basic structural parameters:

nPatches - the number of patches in the model

nStrata - the number of human / host population strata

nHabitats - the number of habitats

xds_setup(
  xds = "xde",
  dlay = "ode",
  MYZname = "RM",
  Xname = "SIS",
  Lname = "trivial",
  nPatches = 1,
  membership = 1,
  residence = 1,
  HPop = 1000,
  Xday = 1,
  MYZday = 1,
  Lday = 1,
  searchB = 1,
  TimeSpent = list(),
  calK = list(),
  searchQ = 1,
  MYZopts = list(),
  Xopts = list(),
  Lopts = list(),
  BFopts = list(),
  model_name = "unnamed"
)

Arguments

xds

is xde/dts for differential / difference equations

dlay

is either "ode" or "dde"

MYZname

is a character string defining a MYZ model

Xname

is a character string defining a X model

Lname

is a character string defining a L model

nPatches

is the number of patches

membership

is a vector that describes the patch where each aquatic habitat is found

residence

is a vector that describes the patch where each human stratum lives

HPop

is the number of humans in each patch

Xday

is the run-time time step for X component (in days): integer or 1/integer

MYZday

is the run-time time step for MYZ component (in days): integer or 1/integer

Lday

is the run-time time step for L component (in days): integer or 1/integer

searchB

is a vector of search weights for blood feeding

TimeSpent

is either a TimeSpent matrix or a string to call a function that sets it up

calK

is either a calK matrix or a string that defines how to set it up

searchQ

is a vector of search weights for egg laying

MYZopts

a list to configure the MYZ model

Xopts

a list to configure the X model

Lopts

a list to configure the L model

BFopts

a list to configure the blood feeding model

model_name

is a name for the model (arbitrary)

Value

a list

Details

This sets up an xds object with pars$frame = 'full'

Parameter values for the models are passed as named lists: Xopts, MYZopts, and Lopts

Search weights, the mosquito dispersal matrix, and the time spent matrix can also be configured using parameters passed at the command line.

Runtime parameters for dts models can be configured by passing values: Xday, MYZday, and Lday

Advanced features must be configured later, including:

multiple-host species or multiple-vector species

exogenous forcing by weather, resources, or other factors

vector control, vaccines, or other mass