Make an xds Object to Study Aquatic Mosquito Ecology

A modified version of xds_setup that streamlines setup for an L Component when the MYZ Component is set to trivial. The model also sets X Component to the trivial module.

The xds object defines frame = class(frame) = 'aquatic' to dispatch xde_derivatives.aquatic or dts_update.aquatic

Usage

xds_setup_aquatic(
  xds = "ode",
  nHabitats = 1,
  Lname = "basicL",
  Lday = 1,
  Lopts = list(),
  MYZopts = list(),
  model_name = "unnamed"
)

Arguments

xds

is ode or dde or dts for ordinary OR delay differential OR difference equations

nHabitats

is the number of habitats

Lname

is a character string defining a L Component module

Lday

is the run-time time step for L Component (in days): integer or 1/integer

Lopts

a list to configure the L Component module

MYZopts

a list to configure F_eggs.trivial

model_name

is a name for the model (arbitrary)

Value

an xds object

See also