Build a Model of Human / Host Epidemiology

A modified version of xds_setup that streamlines setup for models with a trival MY Component.

The xds object defines frame = class(frame) = 'human' to dispatch functions that compute derivatives (xde), update variables (dts), and parse outputs

The MY Component module is set to trivial. The funcion F_fqZ.trivial is called to compute passes the density of infectious adult mosquitoes, and \(f\) and \(q\) can still be configured. In this case, the daily EIR is computed using the blood feeding interface, including Exposure, in the same way as a model with a non-trivial MY Component module.

To study human cohort dynamics by passing a function that computes the daily EIR, consider using xds_setup_eir.

Usage

xds_setup_human(
  Xname = "SIS",
  XHoptions = list(),
  xds = "ode",
  nPatches = 1,
  residence = 1,
  HPop = 1000,
  searchB = 1,
  TimeSpent = list(),
  MYoptions = list(),
  BFopts = list(),
  model_name = "unnamed"
)

Arguments

Xname

a character string defining a X Component module

XHoptions

a named list to configure the X Component module

xds

is ode or dde or dts for ordinary OR delay differential OR difference equations

nPatches

the number of patches

residence

a vector that describes the patch where each human stratum lives

HPop

the number of humans in each patch

searchB

a vector of search weights for blood feeding

TimeSpent

either a TimeSpent matrix or a string to call a function that sets it up

MYoptions

list to configure the MY Component module

BFopts

list to configure the blood feeding model

model_name

a name for the model

Value

an xds model object

See also